refact: findScoreByCafeIdAndMemberId 추가 및 그에 따른 서비스 코드 수정

+++ src/main/java/mocacong/server/dto/response/MyReviewCafeResponse.java
@@ -9,5 +9,5 @@ import lombok.*;
 public class MyReviewCafeResponse {
 
     private String name;
-    private double myScore;
+    private int myScore;
 }
